Fix race condition in flatten
When used by concat, if the current subscription completed at the same
time as a new signal was was received, the new signal could be subscribed to
immediately in addition to the queued up signal that was supposed to be
subscribed to next.

This commit fixes the race condition, making sure that signals are only
subscribed to in the order they are received by the flatten.