Potential fix for a crash
Had an app crash on the line that invokes the subscribeToSignal after
it looked like things were starting to be disposed of. Wondering if
something had been deallocated prior to the subscribeToSignal being set
to nil in the block.

Not sure this will make a difference, but giving it a shot.